refactor: gather the int case of take_from_groups, check GroupedArray indices
`take_from_groups` kept a python loop for int `idx` and stepped slices,
which no caller uses. The int case is now a single gather that raises
IndexError when a group is too short (78 -> 0.8 ms on 50k groups) and
stepped slices raise NotImplementedError. `take` rejects negative group
indices up front instead of failing inside `np.repeat`, and the int32
cast on its result is dropped since it keeps `indptr`'s dtype.

The randomized checks now include empty groups, an empty array, a
horizon longer than the data and out of range int indices.

+++ mlforecast/core.py
@@ -688,7 +688,7 @@ class TimeSeries:
             # self.ga already holds the appended values
             sub = self.ga.take(np.flatnonzero(fresh))
             primed = copy.deepcopy(core)
-            primed.transform(CoreGroupedArray(sub.data, sub.indptr.astype(np.int32)))
+            primed.transform(CoreGroupedArray(sub.data, sub.indptr))
             core.stats_[fresh] = primed.stats_
 
     def _check_aligned_ends(self) -> None:
